Generic udp source / sink
